Moving the nested scrolling behavior back into the SearchLayout, but setting it to offset only the PersistentSearchView instead of the entire layout
