+++ shared/src/main/java/xyz/stream/messenger/shared/view/search/SearchLayout.kt
@@ -9,9 +9,21 @@ import android.view.View
 import android.widget.LinearLayout
 import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.core.content.ContextCompat.getColor
+import androidx.core.math.MathUtils
+import androidx.core.view.ViewCompat
+import androidx.core.view.isVisible
+import androidx.fragment.app.FragmentContainerView
+import androidx.recyclerview.widget.RecyclerView
 import xyz.stream.messenger.shared.R
+import kotlin.math.max
 
-class SearchLayout : CoordinatorLayout {
+class SearchLayout : LinearLayout, CoordinatorLayout.AttachedBehavior {
+
+    private var behavior: Behavior? = null
+
+    private var totalScrollRange = INVALID_SCROLL_RANGE
+    private var downPreScrollRange = INVALID_SCROLL_RANGE
+    private var downScrollRange = INVALID_SCROLL_RANGE
 
     private val searchView: PersistentSearchView by lazy { findViewById<View>(R.id.search_view) as PersistentSearchView }
     private val searchContainer: LinearLayout by lazy { findViewById<View>(R.id.search_container) as LinearLayout }
@@ -22,12 +34,15 @@ class SearchLayout : CoordinatorLayout {
     private var searchViewListener: SearchViewListener? = null
 
     constructor(context: Context, attrs: AttributeSet, defStyleAttr: Int) : super(context, attrs, defStyleAttr) {
+        behavior = Behavior(context, attrs)
     }
 
     constructor(context: Context, attrs: AttributeSet) : super(context, attrs) {
+        behavior = Behavior(context, attrs)
     }
 
     constructor(context: Context) : super(context) {
+        behavior = Behavior()
     }
 
     override fun onAttachedToWindow() {
@@ -90,7 +105,6 @@ class SearchLayout : CoordinatorLayout {
         searchView.swapLeftRightPadding()
         searchViewListener?.onSearchOpened()
         _isSearchOpen = true
-        searchView.isSearchOpen = true
     }
 
     fun closeSearch() {
@@ -105,13 +119,15 @@ class SearchLayout : CoordinatorLayout {
         searchView.text.text = null
         searchViewListener?.onSearchClosed()
         _isSearchOpen = false
-        searchView.isSearchOpen = false
 
         invalidateScrollRanges()
     }
 
     fun invalidateScrollRanges() {
-        searchView.invalidateScrollRanges()
+        searchView.currentOffset = 0
+        totalScrollRange = INVALID_SCROLL_RANGE
+        downPreScrollRange = INVALID_SCROLL_RANGE
+        downScrollRange = INVALID_SCROLL_RANGE
     }
 
     private fun onTextChanged(newText: CharSequence) {
@@ -138,5 +154,148 @@ class SearchLayout : CoordinatorLayout {
     interface SearchViewListener {
         fun onSearchOpened()
         fun onSearchClosed()
+    }
+
+    override fun getBehavior(): CoordinatorLayout.Behavior<*> {
+        return behavior!!
+    }
+
+    internal fun setTopBottomOffset(newOffset: Int, minOffset: Int, maxOffset: Int): Int {
+        var newOffset = newOffset
+        val curOffset: Int = searchView.currentOffset
+        var consumed = 0
+        if (minOffset != 0 && curOffset >= minOffset && curOffset <= maxOffset) {
+            newOffset = MathUtils.clamp(newOffset, minOffset, maxOffset)
+            if (curOffset != newOffset) {
+                searchView.currentOffset = newOffset
+                consumed = curOffset - newOffset
+            }
+        }
+        return consumed
+    }
+
+    internal fun getDownNestedScrollRange(): Int {
+        if (downScrollRange != INVALID_SCROLL_RANGE) {
+            return downScrollRange
+        }
+        var range = 0
+        var i = 0
+        val z = childCount
+        while (i < z) {
+            val child = getChildAt(i)
+            if (child.isVisible) {
+                val lp = child.layoutParams as LinearLayout.LayoutParams
+                var childHeight = child.measuredHeight
+                childHeight += lp.topMargin + lp.bottomMargin
+                range += childHeight
+            }
+            i++
+        }
+        return max(0, range).also { downScrollRange = it }
+    }
+
+    internal fun getDownNestedPreScrollRange(): Int {
+        if (downPreScrollRange != INVALID_SCROLL_RANGE) {
+            return downPreScrollRange
+        }
+        var range = 0
+        for (i in childCount - 1 downTo 0) {
+            val child = getChildAt(i)
+            val childHeight: Int = child.measuredHeight
+            if (range <= 0) {
+                val child = getChildAt(i)
+                if (child.isVisible) {
+                    val lp = child.layoutParams as LinearLayout.LayoutParams
+                    var childRange = lp.topMargin + lp.bottomMargin
+                    childRange += childHeight
+                    range += childRange
+                }
+            } else {
+                break
+            }
+        }
+        return max(0, range).also { downPreScrollRange = it }
+    }
+
+    internal fun getTotalScrollRange(): Int {
+        if (isSearchOpen) return 0
+
+        if (totalScrollRange != INVALID_SCROLL_RANGE) {
+            return totalScrollRange
+        }
+        var range = 0
+        var i = 0
+        val z = childCount
+        while (i < z) {
+            val child = getChildAt(i)
+            if (child.isVisible) {
+                val lp = child.layoutParams as LinearLayout.LayoutParams
+                val childHeight = child.measuredHeight
+                range += childHeight + lp.topMargin + lp.bottomMargin
+            }
+            i++
+        }
+        return max(0, range).also { totalScrollRange = it }
+    }
+
+    class Behavior : CoordinatorLayout.Behavior<SearchLayout> {
+
+        constructor(context: Context, attrs: AttributeSet) : super(context, attrs)
+        constructor() : super()
+
+        override fun layoutDependsOn(parent: CoordinatorLayout, child: SearchLayout, dependency: View): Boolean {
+            return dependency is FragmentContainerView
+        }
+
+        override fun onStartNestedScroll(coordinatorLayout: CoordinatorLayout, child: SearchLayout, directTargetChild: View, target: View, axes: Int, type: Int): Boolean {
+            return axes == ViewCompat.SCROLL_AXIS_VERTICAL ||
+                    super.onStartNestedScroll(coordinatorLayout, child, directTargetChild, target, axes, type)
+        }
+
+        override fun onNestedPreScroll(coordinatorLayout: CoordinatorLayout, child: SearchLayout, target: View, dx: Int, dy: Int, consumed: IntArray, type: Int) {
+            if (target is RecyclerView) {
+                if (!target.canScrollVertically(dy)) {
+                    return
+                }
+            }
+
+            if (dy != 0) {
+                val min: Int
+                val max: Int
+                if (dy < 0) { // We're scrolling down
+                    min = -child.getTotalScrollRange()
+                    max = min + child.getDownNestedPreScrollRange()
+                } else { // We're scrolling up
+                    min = -child.getTotalScrollRange()
+                    max = 0
+                }
+                if (min != max) {
+                    scroll(child, dy, min, max)
+                }
+            }
+        }
+
+        override fun onNestedScroll(coordinatorLayout: CoordinatorLayout, child: SearchLayout, target: View, dxConsumed: Int, dyConsumed: Int, dxUnconsumed: Int, dyUnconsumed: Int, type: Int, consumed: IntArray) {
+            if (dyUnconsumed < 0) {
+                // If the scrolling view is scrolling down but not consuming, it's probably be at
+                // the top of it's content
+                consumed[1] = scroll(child, dyUnconsumed, -child.getDownNestedScrollRange(), 0)
+            }
+        }
+
+        private fun scroll(child: SearchLayout, dy: Int, minOffset: Int, maxOffset: Int): Int {
+            val consumed = child.setTopBottomOffset(
+                    child.searchView.currentOffset - dy,
+                    minOffset,
+                    maxOffset)
+
+
+            ViewCompat.offsetTopAndBottom(child.searchView, -consumed)
+            return consumed
+        }
+    }
+
+    companion object {
+        private const val INVALID_SCROLL_RANGE = -1
     }
 }
